test(core): pin RtcOffset's z channel and local_to_map's rotation sign
Two surviving mutations in `rust/core/src/georef.rs`. Tests only — no
production code changed.

- `RtcOffset::apply` — the z channel could read `self.x` instead of `self.z`
  and nothing failed. `test_rtc_apply` uses `z: 0.0` and never asserts the
  third component of either point, so the whole channel was unconstrained.

- `GeoReference::local_to_map` — flipping the sign in `cos_r * x - sin_r * y`
  survived. The two fixtures between them made it unobservable: one has
  `sin = 1, y = 0`, the other `sin = 0, y != 0`, and in both
  `cos*x - sin*y` and `cos*x + sin*y` are numerically identical. This is the
  sibling of `map_to_local`'s rotation, which was pinned earlier; the
  forward direction had not been.

The `n` term of the same formula was mutated as a control and DIES on main
(1 of 148), so the function is genuinely exercised — the gap is specific to
the `e` term rather than the whole path being untested.

Both are COVERAGE GAPs, not live defects: the production code is correct
today, but nothing failed when it was deliberately broken.

Verified against main directly, with the new tests absent: both mutations
leave 148/148 green there, and both die with the tests in place (1
replacement each, count asserted). A further control on
`RtcOffset::from_positions` died as expected.

Adding the tests pushed `georef.rs` to 841 lines, over its 793-line ratchet
budget, so the tests moved to a sibling `georef_tests.rs` — the pattern
`schema_gen.rs`/`schema_gen_tests.rs` already uses in this crate. No budget
was raised; `georef.rs` is now 713 lines. Single `use super::*;` in the split
file.

Suite: lib 148 -> 150, all 12 test binaries green. Ratchet 4/4. Clippy
reports exactly the same four pre-existing lints and no new ones.
